GetClientRect
GetClientRect returns the co-ordinates of a window's client area - ie. the area of a window minus the title bar and the window's borders.

GetWindowRect returns co-ordinates indicating the area of the whole window, title bar and border included/

This will provide for:-
a more accurate/simpler auto-clicker in DeckExporter
accurate positioning of overlays

This also eliminates the need for both fullscreen and windowed variables in config.xml for the co-ordinates of controls in Hearthstone.