Fix the secrets' tooltips from being 'accessable' after StackPanelSecrets has been changed to hidden.
After all secrets had been triggered and/or game had ended the tooltips were being shown on mouse-over of the area where the secrets panel use to be visible.

Change HideSecrets() to set the secrets panel to 'collapsed' instead of 'hidden'.
Set to 'collapsed' removes the panel from the overlay - UpdateCardTooltip() doesn't show the tooltips on mouse hover-over where the stackpanel existed.