Add config file version tracking
We save a CreatedByVersion string in the config file. With this in
place, we can determine which version of the application wrote the
config file when we load it. This allows us to change the semantics
of config variables in the future and have a mechanism for handling
it without breaking users' legacy config files. We are able to
interpret and update users' legacy config files to current semantics.

If the version doesn't exist in the config file, we assume it was
created prior to this commit.

With this in place, we can remove the hacky handing of when default
window positions were being stored as -32000. We no longer need to
handle this as a special case with the new multi-screen support.
See changes to OpponentWindow.xaml.cs, PlayerWindow.xaml.cs, and
'TimerWindow.xaml.cs'.