Archive decks feature
Add menu items to archive and unarchive decks.

Archived decks will not be visible in the deck picker under 'All' or
the deck's class. They will only be shown under the 'Archived' class
until they are unarchived. The archived class is not shown unless
there is at least one archived deck. Achived decks are marked with
an icon on the deck picker.

Archived decks are automatically unarchived after you edit them or
play a game with them (after a recorded game is assigned to the deck).