Improve setting the minimum window height
In v0.10.0, a bug was introduced where the deckpicker class selector
would scroll down when selected if the main window was at its minimum
height. This required increasing the minimum height due to the
changes in ac125de which increased the size of the deck type selector.

The minimum window height now also updates when the deck search box
is visible. If this causes a change to the height of the window, this
is reversed when the search box is closed or the application is closed.

The minimum window height now also updates when the news status bar is
visible.