half-fix avenge (#1771)
Instead of waiting for more minions to die (relevant when all of them die), don't wait if the count is already zero (i.e. only one minion was on the board). Instead of a likely false positive in for summon deathrattles this should at the worst case result in a false negative, which means the secret triggered and gets removed from the list anyway. (positive: secret greyed out)

Not sure how to fix: two summon-deathrattle minions on the board, getting killed at the same time.