remove zone.log parsing, fix auto deck detection
The only purpose of parsing the zone.log was to detect the heroes. This responsibility is now moved to the power.log.
Deck detection used to break in some cases, because logs would reset to a point after heroes were logged in zone.log, causing Player.Hero and Opponent.Hero to remain empty.