add fallback in-menu detection  (#1919)
In addition to checking for "RegisterFriendChallenge" in Bob.log, HandleInMenu is now also called when prevMode in the LoadingScreen.log is GAMEPLAY. From testing, LoadingScreen seems to log this earlier than Bob, but chances are, this is not consistent (across machines). HandleInMenu can only be called once though, so there should be no issues either way.