fix "quick set tag" not working if too many tags exist (#2127)
The way it's supposed to work: (yes I know this is hacky)
"quick set tag"-checkboxes and SortFilterDecks checkboxes are set to the same binding, but only the latter have an event that updates anything.
So when a QST checkbox changes its state and with that the Tag.Selected property, the one (with the event) also updates and triggers the update.

The problem: The VirtualizingPanel.IsVirtualizing property in ListBoxes is set to True by default, causing any non-visible checkboxes not to load and therefore not to trigger an update.

Disabling VirtualizingPanel.IsVirtualizing fixes this.