constructed stats: use Minutes instead of TotalMinutes for sum/average
Some games seem to have no start or end times set (i.e. it's set to the minimum date value), causing the duration and therefore the total average/sum to be way off. This does not fix that problem entirely but limits the error on a single game to 59 minutes at most. Considering the values are rounded to a one digit anyway, this is good enough.

fixes #2474