Refactor OptionsSearch ahead of supporting more option types
Instead of hard-coding `CheckBox` as the type of option to search for, use a
base class `OptionWrapper<T>`. Currently `CheckBoxWrapper` is the only concrete
child class, but more will be added for #3344.

Use a filter delegate to check whether an element is an "option" so we can
explicitly filter for `CheckBox` types, and for `DockPanel`s that contain a
`CheckBox`. In the future, `DockPanel`s that contain other types of controls
will also be selected. Finding the `DockPanel`s will let us find the related
text depending on the control type, as non-CheckBox types tend to have more
complicated UI.

This change should not affect search results today.