Fix erroneously recording stats for spectator mode
If stats recording for spectator mode games is disabled and an active deck
is selected after spectating a gane (e.g. by starting an arena game), the
spectated game is assigned to the active deck because the end of a 0 turn
game is detected erroneously. This commit prevents the detection of a 0
turn game while in menu.

Fixes #3343.