Fix CapturableOverlay not staying aligned
With the new OverlayBehavior used by BobsBuddyPanel and the
Opponent Board panel, elements are moved "out of" the GridMain
when they are hidden. Since the CapturableOverlay just blindly copies
the Visual of the GridMain, this will lead to these elements staying
visible and the rest of the overlay shifting down.