fix(bobsbuddy): don't hide panel in shopping phase with delayed simulation
Bob's Buddy has a mode where it waits for the end of combat before
performing simulations.

With this commit, to respect user intent, if a user has explicitly
expanded Bob's Buddy during a "Awaiting shopping phase" combat, we now
keep it expanded when transitioning to the shopping phase.
This was previously not the case if the user had the setting to hide
Bob's Buddy during the shopping phase set.

Overall this change should prevent Bob's Buddy frustratingly
disappearing just after you've explicitly expanded it waiting for the
delayed results.