feat: limit memory utilization when parsing large logs
While this will limit the speed with which we can catch up with an existing
Battlegrounds game it should hopefully prevent HDT from running out of memory
where many log lines suddenly appear.

We can further optimize this in future by dynamically updating the cycle time
when HDT is rate-limited.