feat: lazily instantiate DeckEditorViewModel
This would previously happen on app start, regardless of the deck editor
ever being used. Not only does this mean we are needlessly populating
the card database in the editor, but we would also attempt to load card
images. Since this would happen before the card tile downloader was
instantiated we would not see images for these cards, with (currently)
no good way to recover.